+++ Samples/Unpackaged/cpp-console-unpackaged/Unpackaged.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31702.278
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "DynamicDependenciesBasic", "DynamicDependenciesBasic.vcxproj", "{450D819D-68E9-444D-8CB4-ED459B39029A}"
+Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "Unpackaged", "Unpackaged.vcxproj", "{450D819D-68E9-444D-8CB4-ED459B39029A}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
