+++ Samples/Unpackaged/cs-console-unpackaged/Properties/PublishProfiles/FolderProfile.pubxml
