+++ Samples/Unpackaged/cs-console-unpackaged/Unpackaged.csproj
