+++ Samples/Unpackaged/cs-console-unpackaged/Unpackaged.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.31729.503
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DynamicDependenciesBasicCS", "DynamicDependenciesBasicCS.csproj", "{03C525C4-8A8A-4BDD-88A1-10DBD117573C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Unpackaged", "Unpackaged.csproj", "{03C525C4-8A8A-4BDD-88A1-10DBD117573C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
